+++ packages/react/src/context/providers/AssistantRuntimeProvider.tsx
@@ -32,16 +32,14 @@ const useAssistantToolUIsStore = () => {
   return useMemo(() => makeAssistantToolUIsStore(), []);
 };
 
-const useThreadManagerStore = (runtime: AssistantRuntime) => {
-  const [store] = useState(() =>
-    create(() => runtime.threadManager.getState()),
-  );
+const useThreadListStore = (runtime: AssistantRuntime) => {
+  const [store] = useState(() => create(() => runtime.threadList.getState()));
 
   useEffect(() => {
     const updateState = () =>
-      writableStore(store).setState(runtime.threadManager.getState(), true);
+      writableStore(store).setState(runtime.threadList.getState(), true);
     updateState();
-    return runtime.threadManager.subscribe(updateState);
+    return runtime.threadList.subscribe(updateState);
   }, [runtime, store]);
 
   return store;
@@ -52,15 +50,15 @@ export const AssistantRuntimeProviderImpl: FC<
 > = ({ children, runtime }) => {
   const useAssistantRuntime = useAssistantRuntimeStore(runtime);
   const useToolUIs = useAssistantToolUIsStore();
-  const useThreadManager = useThreadManagerStore(runtime);
+  const useThreadList = useThreadListStore(runtime);
   const context = useMemo(() => {
     return {
       useToolUIs,
       useAssistantRuntime,
       useAssistantActions: useAssistantRuntime,
-      useThreadManager,
+      useThreadList,
     };
-  }, [useAssistantRuntime, useToolUIs, useThreadManager]);
+  }, [useAssistantRuntime, useToolUIs, useThreadList]);
 
   return (
     <AssistantContext.Provider value={context}>